pyproject.toml
[project]
name = "iedb"
version = "2.0.0"
description = "IEDB - Intelligent Enterprise Database"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "Niranjoy Yengkhom", email = "niranjoyy@gmail.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Security :: Cryptography",
]
keywords = ["database", "blockchain", "encryption", "jwt", "authentication", "enterprise"]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
    "cryptography>=41.0.7",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
]
packaging = [
    "pyinstaller>=6.0.0",
]
all = [
    "iedb[dev,packaging]",
]

[project.urls]
"Homepage" = "https://github.com/Niranjoyyengkhom/Blockdb-1.2"
"Bug Tracker" = "https://github.com/Niranjoyyengkhom/Blockdb-1.2/issues"
"Documentation" = "https://github.com/Niranjoyyengkhom/Blockdb-1.2/wiki"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["iedb"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
iedb = ["py.typed", "**/*.json", "**/*.html", "**/*.css", "**/*.js"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
